requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
pymongo==4.8.0
python-dotenv==1.0.1
pydantic-settings==2.11.0
streamlit==1.39.0

# LangChain and LangGraph dependencies
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.31
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langgraph==0.2.59
langsmith==0.4.37
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1
python-docx==1.1.2

# Data analysis and visualization
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
jupyter==1.0.0
ipykernel==6.29.4

# DSPy for optimized LLM applications
dspy-ai==2.5.29

# Testing dependencies
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-mock==3.14.0

